+++ lib/cards_systems.rb
@@ -1,4 +1,11 @@
 require 'cards_systems/version'
+require 'cards_systems/support'
+require 'cards_systems/menu'
+require 'cards_systems/command/delete'
+require 'cards_systems/command/insert'
+require 'cards_systems/command/quit'
+require 'cards_systems/command/search'
+require 'cards_systems/command/show'
 require 'cards_systems/application'
 
 autoload :Card,            'cards_systems/card'
