+++ lib/cards_systems/command/quit.rb
@@ -1,0 +1,7 @@
+module Command
+  class Quit
+    def run
+      exit 0
+    end
+  end
+end
