+++ lib/cards_systems/command/show.rb
@@ -1,0 +1,9 @@
+module Command
+  class Show
+    include Support
+
+    def run
+      show_cards(Card.all)
+    end
+  end
+end
